Feat: api, custom hook, scroll event 처리 index.tsx에서 하던거 분리
